+++ docs/fundamentals/runtime-libraries/snippets/System.Diagnostics/PerformanceCounterType/Overview/csharp/numberofitems64.cs
@@ -1,17 +1,17 @@
 ﻿//<snippet1>
 using System;
 using System.Collections;
-using System.Collections.Specialized;
 using System.Diagnostics;
+using System.Runtime.Versioning;
 
-public class NumberOfItems64
+[SupportedOSPlatform("Windows")]
+public class NumberOfItems64_1
 {
+    private static PerformanceCounter PC;
 
-	private static PerformanceCounter PC;
-
-	public static void Main()
-	{	
-		ArrayList samplesList = new ArrayList();
+    public static void Main()
+    {
+        ArrayList samplesList = new ArrayList();
 
         // If the category does not exist, create the category and exit.
         // Perfomance counters should not be created and immediately used.
@@ -20,36 +20,35 @@ public class NumberOfItems64
         // Execute this sample a second time to use the category.
         if (SetupCategory())
             return;
-		CreateCounters();
-		CollectSamples(samplesList);
-		CalculateResults(samplesList);
-	}
+        CreateCounters();
+        CollectSamples(samplesList);
+        CalculateResults(samplesList);
+    }
 
-	private static bool SetupCategory()
-	{		
-		if ( !PerformanceCounterCategory.Exists("NumberOfItems64SampleCategory") )
-		{
+    private static bool SetupCategory()
+    {
+        if (!PerformanceCounterCategory.Exists("NumberOfItems64SampleCategory"))
+        {
+            CounterCreationDataCollection CCDC = new CounterCreationDataCollection();
 
-			CounterCreationDataCollection CCDC = new CounterCreationDataCollection();
+            // Add the counter.
+            CounterCreationData NOI64 = new CounterCreationData();
+            NOI64.CounterType = PerformanceCounterType.NumberOfItems64;
+            NOI64.CounterName = "NumberOfItems64Sample";
+            CCDC.Add(NOI64);
 
-			// Add the counter.
-			CounterCreationData NOI64 = new CounterCreationData();
-			NOI64.CounterType = PerformanceCounterType.NumberOfItems64;
-			NOI64.CounterName = "NumberOfItems64Sample";
-			CCDC.Add(NOI64);
-
-			// Create the category.
-			PerformanceCounterCategory.Create("NumberOfItems64SampleCategory",
+            // Create the category.
+            PerformanceCounterCategory.Create("NumberOfItems64SampleCategory",
                 "Demonstrates usage of the NumberOfItems64 performance counter type.",
                 PerformanceCounterCategoryType.SingleInstance, CCDC);
-			return(true);
-		}
-		else
-		{
-			Console.WriteLine("Category exists - NumberOfItems64SampleCategory");
-			return(false);
-		}
-	}
+            return (true);
+        }
+        else
+        {
+            Console.WriteLine("Category exists - NumberOfItems64SampleCategory");
+            return (false);
+        }
+    }
 
     private static void CreateCounters()
     {
@@ -58,18 +57,17 @@ public class NumberOfItems64
             "NumberOfItems64Sample",
             false);
 
-        PC.RawValue=0;
+        PC.RawValue = 0;
     }
 
     private static void CollectSamples(ArrayList samplesList)
     {
-		
-        Random r = new Random( DateTime.Now.Millisecond );
+        Random r = new Random(DateTime.Now.Millisecond);
 
         // Loop for the samples.
         for (int j = 0; j < 100; j++)
         {
-	
+
             int value = r.Next(1, 10);
             Console.Write(j + " = " + value);
 
@@ -78,7 +76,7 @@ public class NumberOfItems64
             if ((j % 10) == 9)
             {
                 OutputSample(PC.NextSample());
-                samplesList.Add( PC.NextSample() );
+                samplesList.Add(PC.NextSample());
             }
             else
             {
@@ -89,48 +87,48 @@ public class NumberOfItems64
         }
     }
 
-	private static void CalculateResults(ArrayList samplesList)
-	{
-		for(int i = 0; i < (samplesList.Count - 1); i++)
-		{
-			// Output the sample.
-			OutputSample( (CounterSample)samplesList[i] );
-			OutputSample( (CounterSample)samplesList[i+1] );
+    private static void CalculateResults(ArrayList samplesList)
+    {
+        for (int i = 0; i < (samplesList.Count - 1); i++)
+        {
+            // Output the sample.
+            OutputSample((CounterSample)samplesList[i]);
+            OutputSample((CounterSample)samplesList[i + 1]);
 
             // Use .NET to calculate the counter value.
-			Console.WriteLine(".NET computed counter value = " +
-				CounterSampleCalculator.ComputeCounterValue((CounterSample)samplesList[i],
-				(CounterSample)samplesList[i+1]) );
+            Console.WriteLine(".NET computed counter value = " +
+                CounterSampleCalculator.ComputeCounterValue((CounterSample)samplesList[i],
+                (CounterSample)samplesList[i + 1]));
 
             // Calculate the counter value manually.
-			Console.WriteLine("My computed counter value = " +
-				MyComputeCounterValue((CounterSample)samplesList[i],
-				(CounterSample)samplesList[i+1]) );
-		}
-	}
+            Console.WriteLine("My computed counter value = " +
+                MyComputeCounterValue((CounterSample)samplesList[i],
+                (CounterSample)samplesList[i + 1]));
+        }
+    }
 
-	//++++++++//++++++++//++++++++//++++++++//++++++++//++++++++//++++++++//++++++++
-	//++++++++//++++++++//++++++++//++++++++//++++++++//++++++++//++++++++//++++++++
-	private static Single MyComputeCounterValue(CounterSample s0, CounterSample s1)
-	{
-		Single counterValue = s1.RawValue;
-		return(counterValue);
-	}
-	
-	// Output information about the counter sample.
+    //++++++++//++++++++//++++++++//++++++++//++++++++//++++++++//++++++++//++++++++
+    //++++++++//++++++++//++++++++//++++++++//++++++++//++++++++//++++++++//++++++++
+    private static Single MyComputeCounterValue(CounterSample s0, CounterSample s1)
+    {
+        Single counterValue = s1.RawValue;
+        return (counterValue);
+    }
+
+    // Output information about the counter sample.
     private static void OutputSample(CounterSample s)
-	{
-		Console.WriteLine("\r\n+++++++++++");
-		Console.WriteLine("Sample values - \r\n");
-		Console.WriteLine("   BaseValue        = " + s.BaseValue);
-		Console.WriteLine("   CounterFrequency = " + s.CounterFrequency);
-		Console.WriteLine("   CounterTimeStamp = " + s.CounterTimeStamp);
-		Console.WriteLine("   CounterType      = " + s.CounterType);
-		Console.WriteLine("   RawValue         = " + s.RawValue);
-		Console.WriteLine("   SystemFrequency  = " + s.SystemFrequency);
-		Console.WriteLine("   TimeStamp        = " + s.TimeStamp);
-		Console.WriteLine("   TimeStamp100nSec = " + s.TimeStamp100nSec);
-		Console.WriteLine("++++++++++++++++++++++");
-	}
+    {
+        Console.WriteLine("\r\n+++++++++++");
+        Console.WriteLine("Sample values - \r\n");
+        Console.WriteLine("   BaseValue        = " + s.BaseValue);
+        Console.WriteLine("   CounterFrequency = " + s.CounterFrequency);
+        Console.WriteLine("   CounterTimeStamp = " + s.CounterTimeStamp);
+        Console.WriteLine("   CounterType      = " + s.CounterType);
+        Console.WriteLine("   RawValue         = " + s.RawValue);
+        Console.WriteLine("   SystemFrequency  = " + s.SystemFrequency);
+        Console.WriteLine("   TimeStamp        = " + s.TimeStamp);
+        Console.WriteLine("   TimeStamp100nSec = " + s.TimeStamp100nSec);
+        Console.WriteLine("++++++++++++++++++++++");
+    }
 }
 //</snippet1>
