+++ docs/fundamentals/runtime-libraries/snippets/System.Diagnostics/PerformanceCounterType/Overview/csharp/rateofcountspersecond32.cs
@@ -1,16 +1,17 @@
 ﻿//<snippet1>
 using System;
 using System.Collections;
-using System.Collections.Specialized;
 using System.Diagnostics;
+using System.Runtime.Versioning;
 
-public class App
+[SupportedOSPlatform("Windows")]
+public class App3
 {
     private static PerformanceCounter PC;
 
-	public static void Main()
-	{	
-		ArrayList samplesList = new ArrayList();
+    public static void Main()
+    {
+        ArrayList samplesList = new ArrayList();
 
         // If the category does not exist, create the category and exit.
         // Perfomance counters should not be created and immediately used.
@@ -20,14 +21,13 @@ public class App
         if (SetupCategory())
             return;
         CreateCounters();
-		CollectSamples(samplesList);
-		CalculateResults(samplesList);
-	}
+        CollectSamples(samplesList);
+        CalculateResults(samplesList);
+    }
 
     private static bool SetupCategory()
     {
-		
-        if ( !PerformanceCounterCategory.Exists("RateOfCountsPerSecond32SampleCategory") )
+        if (!PerformanceCounterCategory.Exists("RateOfCountsPerSecond32SampleCategory"))
         {
 
             CounterCreationDataCollection CCDC = new CounterCreationDataCollection();
@@ -37,17 +37,17 @@ public class App
             rateOfCounts32.CounterType = PerformanceCounterType.RateOfCountsPerSecond32;
             rateOfCounts32.CounterName = "RateOfCountsPerSecond32Sample";
             CCDC.Add(rateOfCounts32);
-	
-             // Create the category.
+
+            // Create the category.
             PerformanceCounterCategory.Create("RateOfCountsPerSecond32SampleCategory",
                 "Demonstrates usage of the RateOfCountsPerSecond32 performance counter type.",
                 PerformanceCounterCategoryType.SingleInstance, CCDC);
-              return(true);
+            return (true);
         }
         else
         {
             Console.WriteLine("Category exists - RateOfCountsPerSecond32SampleCategory");
-            return(false);
+            return (false);
         }
     }
 
@@ -58,13 +58,13 @@ public class App
             "RateOfCountsPerSecond32Sample",
             false);
 
-        PC.RawValue=0;
+        PC.RawValue = 0;
     }
 
     private static void CollectSamples(ArrayList samplesList)
     {
-	
-        Random r = new Random( DateTime.Now.Millisecond );
+
+        Random r = new Random(DateTime.Now.Millisecond);
 
         // Initialize the performance counter.
         PC.NextSample();
@@ -72,7 +72,7 @@ public class App
         // Loop for the samples.
         for (int j = 0; j < 100; j++)
         {
-	
+
             int value = r.Next(1, 10);
             PC.IncrementBy(value);
             Console.Write(j + " = " + value);
@@ -81,7 +81,7 @@ public class App
             {
                 Console.WriteLine(";       NextValue() = " + PC.NextValue().ToString());
                 OutputSample(PC.NextSample());
-                samplesList.Add( PC.NextSample() );
+                samplesList.Add(PC.NextSample());
             }
             else
             {
@@ -92,68 +92,68 @@ public class App
         }
     }
 
-	private static void CalculateResults(ArrayList samplesList)
-	{
-		for(int i = 0; i < (samplesList.Count - 1); i++)
-		{
-			// Output the sample.
-			OutputSample( (CounterSample)samplesList[i] );
-			OutputSample( (CounterSample)samplesList[i+1] );
+    private static void CalculateResults(ArrayList samplesList)
+    {
+        for (int i = 0; i < (samplesList.Count - 1); i++)
+        {
+            // Output the sample.
+            OutputSample((CounterSample)samplesList[i]);
+            OutputSample((CounterSample)samplesList[i + 1]);
 
             // Use .NET to calculate the counter value.
-			Console.WriteLine(".NET computed counter value = " +
-				CounterSampleCalculator.ComputeCounterValue((CounterSample)samplesList[i],
-				(CounterSample)samplesList[i+1]) );
+            Console.WriteLine(".NET computed counter value = " +
+                CounterSampleCalculator.ComputeCounterValue((CounterSample)samplesList[i],
+                (CounterSample)samplesList[i + 1]));
 
             // Calculate the counter value manually.
             Console.WriteLine("My computed counter value = " +
-				MyComputeCounterValue((CounterSample)samplesList[i],
-				(CounterSample)samplesList[i+1]) );
-		}
-	}
+                MyComputeCounterValue((CounterSample)samplesList[i],
+                (CounterSample)samplesList[i + 1]));
+        }
+    }
 
-	//++++++++//++++++++//++++++++//++++++++//++++++++//++++++++//++++++++//++++++++
-	//	PERF_COUNTER_COUNTER
-	//	Description	 - This counter type shows the average number of operations completed
-	//		during each second of the sample interval. Counters of this type
-	//		measure time in ticks of the system clock. The F variable represents
-	//		the number of ticks per second. The value of F is factored into the
-	//		equation so that the result can be displayed in seconds.
-	//
+    //++++++++//++++++++//++++++++//++++++++//++++++++//++++++++//++++++++//++++++++
+    //	PERF_COUNTER_COUNTER
+    //	Description	 - This counter type shows the average number of operations completed
+    //		during each second of the sample interval. Counters of this type
+    //		measure time in ticks of the system clock. The F variable represents
+    //		the number of ticks per second. The value of F is factored into the
+    //		equation so that the result can be displayed in seconds.
+    //
     //	Generic type - Difference
-	//
-	//	Formula - (N1 - N0) / ( (D1 - D0) / F), where the numerator (N) represents the number
-	//		of operations performed during the last sample interval, the denominator
-	//		(D) represents the number of ticks elapsed during the last sample
-	//		interval, and F is the frequency of the ticks.
-	//
-	//	     Average - (Nx - N0) / ((Dx - D0) / F)
-	//
-	//       Example - System\ File Read Operations/sec
-	//++++++++//++++++++//++++++++//++++++++//++++++++//++++++++//++++++++//++++++++
-	private static Single MyComputeCounterValue(CounterSample s0, CounterSample s1)
-	{
-		Single numerator = (Single)(s1.RawValue - s0.RawValue);
-		Single denomenator = (Single)(s1.TimeStamp - s0.TimeStamp) / (Single)s1.SystemFrequency;
-		Single counterValue = numerator / denomenator;
-		return(counterValue);
-	}
-	
+    //
+    //	Formula - (N1 - N0) / ( (D1 - D0) / F), where the numerator (N) represents the number
+    //		of operations performed during the last sample interval, the denominator
+    //		(D) represents the number of ticks elapsed during the last sample
+    //		interval, and F is the frequency of the ticks.
+    //
+    //	     Average - (Nx - N0) / ((Dx - D0) / F)
+    //
+    //       Example - System\ File Read Operations/sec
+    //++++++++//++++++++//++++++++//++++++++//++++++++//++++++++//++++++++//++++++++
+    private static Single MyComputeCounterValue(CounterSample s0, CounterSample s1)
+    {
+        Single numerator = (Single)(s1.RawValue - s0.RawValue);
+        Single denomenator = (Single)(s1.TimeStamp - s0.TimeStamp) / (Single)s1.SystemFrequency;
+        Single counterValue = numerator / denomenator;
+        return (counterValue);
+    }
+
     // Output information about the counter sample.
-	private static void OutputSample(CounterSample s)
-	{
-		Console.WriteLine("\r\n+++++++++++");
-		Console.WriteLine("Sample values - \r\n");
-		Console.WriteLine("   BaseValue        = " + s.BaseValue);
-		Console.WriteLine("   CounterFrequency = " + s.CounterFrequency);
-		Console.WriteLine("   CounterTimeStamp = " + s.CounterTimeStamp);
-		Console.WriteLine("   CounterType      = " + s.CounterType);
-		Console.WriteLine("   RawValue         = " + s.RawValue);
-		Console.WriteLine("   SystemFrequency  = " + s.SystemFrequency);
-		Console.WriteLine("   TimeStamp        = " + s.TimeStamp);
-		Console.WriteLine("   TimeStamp100nSec = " + s.TimeStamp100nSec);
-		Console.WriteLine("++++++++++++++++++++++");
-	}
+    private static void OutputSample(CounterSample s)
+    {
+        Console.WriteLine("\r\n+++++++++++");
+        Console.WriteLine("Sample values - \r\n");
+        Console.WriteLine("   BaseValue        = " + s.BaseValue);
+        Console.WriteLine("   CounterFrequency = " + s.CounterFrequency);
+        Console.WriteLine("   CounterTimeStamp = " + s.CounterTimeStamp);
+        Console.WriteLine("   CounterType      = " + s.CounterType);
+        Console.WriteLine("   RawValue         = " + s.RawValue);
+        Console.WriteLine("   SystemFrequency  = " + s.SystemFrequency);
+        Console.WriteLine("   TimeStamp        = " + s.TimeStamp);
+        Console.WriteLine("   TimeStamp100nSec = " + s.TimeStamp100nSec);
+        Console.WriteLine("++++++++++++++++++++++");
+    }
 }
 
 //</snippet1>
