+++ docs/fundamentals/runtime-libraries/snippets/System.Diagnostics/PerformanceCounterType/Overview/csharp/rateofcountspersecond64.cs
@@ -1,10 +1,11 @@
 ﻿//<snippet1>
 using System;
 using System.Collections;
-using System.Collections.Specialized;
 using System.Diagnostics;
+using System.Runtime.Versioning;
 
-public class App
+[SupportedOSPlatform("Windows")]
+public class App4
 {
     private static PerformanceCounter PC;
